separate sqlite concerns
